+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2.0 (unreleased)
 
 - Added support for models with external data
+- Added `device` option
 
 ## 1.1.1 (2024-10-14)